utils/llm_engine.py
import os
import json
from pydantic import BaseModel, Field
from typing import List, Optional
from google import genai
from google.genai import types

class AuditResult(BaseModel):
    classification: str = Field(description="Must be 'Marketing Fluff' or 'Evidence-Based'")
    buzzwords: List[str] = Field(description="List of detected greenwashing or sustainability terms")
    score: int = Field(description="Sustainability score from 0 to 100")
    explanation: str = Field(description="A 1-2 sentence explanation of the score and classification")

FEW_SHOT_PROMPT = """You are an expert Sustainability Auditor.

Your task is to analyze product descriptions and classify them as:
- "Marketing Fluff" (vague, unverifiable greenwashing)
- "Evidence-Based" (specific, measurable, certified claims)

You MUST follow a hybrid reasoning approach:
1. Detect vague buzzwords
2. Detect strong signals (certifications, numbers, standards)

Definitions:
- Vague words: eco-friendly, green, natural, sustainable, organic, conscious
- Strong signals: certifications (GOTS, Fair Trade, B-Corp), percentages, recycled materials, carbon data

Real-world dataset examples:
Example A:
"Eco-friendly lifestyle product designed for conscious consumers."
→ Marketing Fluff

Example B:
"Made with 70% recycled polyester and certified by Global Recycled Standard."
→ Evidence-Based

Scoring Guidelines:
- Penalize vague claims
- Reward certifications and measurable facts
- Final score must be between 0-100

Return ONLY valid JSON matching exactly the requested schema.

Now analyze:
Text: {user_input}
"""

def analyze_text(text: str) -> Optional[dict]:
    api_key = os.getenv("GEMINI_API_KEY")
    result_dict = None
    
    # 🚨 HACKATHON FALLBACK: If key is invalid or missing, return a smart mock response so UI keeps working!
    if not api_key or api_key == "your_api_key_here":
        if "certified" in text.lower() or "recycled" in text.lower():
            result_dict = {
                "classification": "Evidence-Based",
                "buzzwords": ["certified", "biodegradable", "recyclable", "post consumer recycled plastic", "renewable energy"],
                "explanation": "The claims are backed by certifications and specific, measurable materials usage."
            }
        else:
            result_dict = {
                "classification": "Marketing Fluff",
                "buzzwords": ["eco-friendly", "green", "natural", "multivitamin", "essential"],
                "explanation": "Uses vague assertions without any verifiable data or recognized third-party certifications."
            }
    else:
        client = genai.Client(api_key=api_key)
        prompt = FEW_SHOT_PROMPT.replace("{user_input}", text)
        
        try:
            response = client.models.generate_content(
                model='gemini-2.5-flash',
                contents=prompt,
                config=types.GenerateContentConfig(
                    response_mime_type="application/json",
                    response_schema=AuditResult,
                ),
            )
            result_dict = json.loads(response.text)
        except Exception as e:
            print(f"Error during LLM API call: {e}")
            if "certified" in text.lower() or "recycled" in text.lower():
                 result_dict = {
                     "classification": "Evidence-Based",
                     "buzzwords": ["certified", "biodegradable", "recyclable", "sustainable"],
                     "explanation": "[Mock Fallback] Details specific certifications and material stats."
                 }
            else:
                result_dict = {
                    "classification": "Marketing Fluff",
                    "buzzwords": ["natural", "eco-friendly", "immunity"],
                     "explanation": "[Mock Fallback] The claims rely on vague marketing terms."
                }
                
    if result_dict:
        # Step 1: Track components
        breakdown = {
            "certifications": 0,
            "measurable_data": 0,
            "transparency": 0,
            "vague_claims": 0,
            "lack_of_evidence": 0
        }
        
        text_lower = text.lower()
        
        # Step 2: Fill it during scoring
        vague_terms = ["eco-friendly", "green", "natural", "sustainable", "organic", "conscious", "overall health", "fluff"]
        
        # --- RAG-Lite Verification ---
        cert_db = {}
        try:
            with open(os.path.join(os.path.dirname(os.path.dirname(__file__)), "data", "certifications.json"), "r") as f:
                cert_db = json.load(f)
        except Exception:
            pass # fallback
            
        verified_claims = []
        for cert, definition in cert_db.items():
            if cert.lower() in text_lower:
                verified_claims.append(f"{cert}: {definition}")
        
        vague_found = any(word in text_lower for word in vague_terms) or "Fluff" in result_dict.get("classification", "")
        strong_found = len(verified_claims) > 0
        has_numbers = any(char.isdigit() for char in text)
        
        if vague_found:
            breakdown["vague_claims"] = -20
            
        if not strong_found and not has_numbers:
            breakdown["lack_of_evidence"] = -30
            
        if has_numbers:
            breakdown["measurable_data"] = 20
            
        if strong_found:
            breakdown["certifications"] = 45 # Massive RAG boost!
            result_dict["explanation"] = " Verified Certifications Found: " + " | ".join(verified_claims)
            
        if "ingredient" in text_lower or "material" in text_lower or "supply chain" in text_lower:
            breakdown["transparency"] = 10
            
        final_score = 50 + sum(breakdown.values())
        final_score = max(0, min(100, final_score))
        
        result_dict["trust_breakdown"] = breakdown
        result_dict["score"] = final_score
        
    return result_dict
