requirements.txt
streamlit
google-genai
trafilatura
datasets
python-dotenv
pydantic
lxml_html_clean
